Moved MP tracing to microprofile module.
