+++ bom/pom.xml
@@ -313,8 +313,8 @@
             </dependency>
             <dependency>
                 <!-- MP extension -->
-                <groupId>io.helidon.tracing</groupId>
-                <artifactId>helidon-tracing-microprofile</artifactId>
+                <groupId>io.helidon.microprofile.tracing</groupId>
+                <artifactId>helidon-microprofile-tracing</artifactId>
                 <version>${project.version}</version>
             </dependency>
