+++ microprofile/tracing/src/main/resources/META-INF/beans.xml
