--- tracing/microprofile/pom.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.helidon.tracing</groupId>
-        <artifactId>helidon-tracing-project</artifactId>
-        <version>0.11.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>helidon-tracing-microprofile</artifactId>
-    <name>Helidon Tracing Microprofile</name>
-
-    <description>
-        Support for distributed tracing in Helidon MP
-        and
-        TODO implementation of Microprofile Opentracing specification.
-    </description>
-
-    <dependencies>
-        <!--<dependency>-->
-            <!--<groupId>org.eclipse.microprofile.opentracing</groupId>-->
-            <!--<artifactId>microprofile-opentracing-api</artifactId>-->
-            <!--&lt;!&ndash; TODO move to parent &ndash;&gt;-->
-            <!--<version>1.2.1</version>-->
-        <!--</dependency>-->
-        <dependency>
-            <groupId>io.helidon.tracing</groupId>
-            <artifactId>helidon-tracing-jersey</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.helidon.microprofile.server</groupId>
-            <artifactId>helidon-microprofile-server</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.helidon.microprofile.config</groupId>
-            <artifactId>helidon-microprofile-config-cdi</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.helidon.microprofile.bundles</groupId>
-            <artifactId>internal-test-libs</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-</project>
