--- tracing/microprofile/src/main/java/io/helidon/tracing/microprofile/MpTracingAutoDiscoverable.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.tracing.microprofile;
-
-import javax.ws.rs.ConstrainedTo;
-import javax.ws.rs.RuntimeType;
-import javax.ws.rs.core.FeatureContext;
-
-import org.glassfish.jersey.internal.spi.AutoDiscoverable;
-
-/**
- * Registers the {@link MpTracingContextFilter} to support
- *  propagation of information from server runtime to client runtime.
- */
-@ConstrainedTo(RuntimeType.SERVER)
-public class MpTracingAutoDiscoverable implements AutoDiscoverable {
-    @Override
-    public void configure(FeatureContext context) {
-        context.register(MpTracingContextFilter.class);
-    }
-}
